Move `useImageSelection`
