+++ apps/website/src/hooks/useImageSelection.ts
